+++ src/lib/pages/contracts/data.ts
@@ -1,5 +1,6 @@
 import type { ContractInfo } from "lib/types";
 
+import { useQueryEvents } from "lib/hooks";
 import { useContractStore } from "lib/providers/store";
 import { useContracts } from "lib/services/wasm/contract";
 
@@ -9,14 +10,16 @@ export const useRecentContracts = (
   setTotalData: (totalData: number) => void
 ) => {
   const { getContractLocalInfo } = useContractStore();
-  const { data: contracts, isLoading } = useContracts(pageSize, offset, {
+  const contractQuery = useContracts(pageSize, offset);
+  useQueryEvents(contractQuery, {
     onSuccess: (data) => setTotalData(data.total),
   });
+  const { data: contracts, isLoading } = contractQuery;
 
   if (!contracts) return { data: undefined, isLoading };
   return {
     data: {
-      items: contracts.items.map<ContractInfo>((contract) => {
+      items: contractQuery.data.items.map<ContractInfo>((contract) => {
         const localInfo = getContractLocalInfo(contract.contractAddress);
         return {
           ...contract,